requirements.txt
transformers==4.0.1
hydra-core==1.1.0
torch==1.6.0
Flask==1.1.2